refactor: versioning require
Simplify when statically selected
