+++ lib/config/definitions.js
@@ -1362,7 +1362,7 @@ const options = [
     ],
     cli: false,
     env: false,
-  }
+  },
 ];
 
 function getOptions() {
