+++ lib/datasource/npm.js
@@ -8,7 +8,7 @@ const getRegistryUrl = require('registry-auth-token/registry-url');
 const registryAuthToken = require('registry-auth-token');
 const parse = require('github-url-from-git');
 const { isBase64 } = require('validator');
-const { isVersion, sortVersions } = require('../versioning')('semver');
+const { isVersion, sortVersions } = require('../versioning/semver');
 const hostRules = require('../util/host-rules');
 
 module.exports = {
