+++ lib/datasource/packagist.js
@@ -4,7 +4,7 @@ const delay = require('delay');
 const got = require('got');
 const parse = require('github-url-from-git');
 const pAll = require('p-all');
-const { isVersion, sortVersions } = require('../versioning')('semverComposer');
+const { isVersion, sortVersions } = require('../versioning/semver-composer');
 const hostRules = require('../util/host-rules');
 
 module.exports = {
