+++ lib/datasource/pypi.js
@@ -1,7 +1,7 @@
 const got = require('got');
 const url = require('url');
 const is = require('@sindresorhus/is');
-const { isVersion, sortVersions, matches } = require('../versioning')('pep440');
+const { isVersion, sortVersions, matches } = require('../versioning/pep440');
 
 module.exports = {
   getPkgReleases,
