+++ lib/datasource/terraform.js
@@ -1,6 +1,6 @@
 const got = require('got');
 const parse = require('github-url-from-git');
-const { isVersion, sortVersions } = require('../versioning')('semver');
+const { isVersion, sortVersions } = require('../versioning/semver');
 
 module.exports = {
   getPkgReleases,
