+++ lib/manager/buildkite/extract.js
@@ -1,6 +1,4 @@
-const versioning = require('../../versioning');
-
-const { isVersion } = versioning('semver');
+const { isVersion } = require('../../versioning/semver');
 
 module.exports = {
   extractPackageFile,
