+++ lib/manager/composer/extract.js
@@ -1,5 +1,5 @@
 const is = require('@sindresorhus/is');
-const semverComposer = require('../../versioning')('semverComposer');
+const semverComposer = require('../../versioning/semver-composer');
 
 module.exports = {
   extractPackageFile,
