+++ lib/manager/npm/extract/index.js
@@ -3,9 +3,7 @@ const upath = require('upath');
 const { getLockedVersions } = require('./locked-versions');
 const { detectMonorepos } = require('./monorepo');
 const { mightBeABrowserLibrary } = require('./type');
-const versioning = require('../../../versioning');
-
-const semver = versioning('semver');
+const semver = require('../../../versioning/semver');
 
 module.exports = {
   extractAllPackageFiles,
