+++ lib/manager/travis/package.js
@@ -1,9 +1,7 @@
 const is = require('@sindresorhus/is');
 const { isEqual } = require('lodash');
 const { getPkgReleases } = require('../../datasource');
-const versioning = require('../../versioning');
-
-const { isVersion, maxSatisfyingVersion } = versioning('semver');
+const { isVersion, maxSatisfyingVersion } = require('../../versioning/semver');
 
 module.exports = {
   getPackageUpdates,
